Upgrade Clap
- Use name to reference flags with dashes
- Switch the verbose to be a u8
- OS string parsing works by default now without needing to be parsed

+++ cargo-cyclonedx/src/main.rs
@@ -110,7 +110,7 @@ fn setup_logging(args: &Args, config: &mut Config) -> anyhow::Result<()> {
 
     // configure logging level of cargo to match what was passed via CLI
     config.configure(
-        args.verbose,
+        args.verbose as u32,
         args.quiet,
         None,
         false,
